review changes, adding agenttraceconfigurator
